requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
cryptography==46.0.7
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.128.8
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
groq==1.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
narwhals==2.19.0
numpy==2.0.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
postgrest==2.28.3
propcache==0.4.1
protobuf==6.33.6
pyarrow==21.0.0
pycparser==2.23
pydantic==2.13.0
pydantic_core==2.46.0
pydeck==0.9.1
Pygments==2.20.0
pyiceberg==0.10.0
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
realtime==2.28.3
referencing==0.36.2
requests==2.32.5
rich==14.3.4
rpds-py==0.27.1
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.49.3
storage3==2.28.3
streamlit==1.50.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
tenacity==9.1.2
toml==0.10.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.39.0
watchdog==6.0.0
websockets==15.0.1
yarl==1.22.0
